+++ src/main/java/org/prgms/locomocoserver/chat/application/MongoChatMessageService.java
@@ -10,7 +10,7 @@ import org.prgms.locomocoserver.chat.dto.ChatUserInfo;
 import org.prgms.locomocoserver.chat.dto.request.ChatMessageRequestDto;
 import org.prgms.locomocoserver.chat.exception.ChatErrorType;
 import org.prgms.locomocoserver.chat.exception.ChatException;
-import org.prgms.locomocoserver.chat.querydsl.ChatRoomCustomRepository;
+import org.prgms.locomocoserver.chat.domain.querydsl.ChatRoomCustomRepository;
 import org.prgms.locomocoserver.user.application.UserService;
 import org.prgms.locomocoserver.user.domain.User;
 import org.prgms.locomocoserver.user.exception.UserException;
